requirements.txt
python-dotenv
keyboard
mouse
Pillow
